Provide defaults to "parse_command_line".
Rationale for the current implementation: using an explicit merge of
the options provided in the config file and command line options
allows us to swap the order in which "load_config_file" and
"parse_command_line" are called. That way switches provided in the
command line may influence the processing of the config file.

Using meaningful defaults in "parse_command_line" we loose the ability
to detect if a switch has been provided or is using the default,
effectively imposing an order to the calls.

This flexibility is not required for "z80count" so we may loose it in
exchange of a simpler "parse_command_line". Reverting to the current
implementation would be a matter of passing an _all None_ dictionary
to "parse_command_line" and adding back the explicit merge.